% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLMing-package.R
\docType{package}
\name{LLMing-package}
\alias{LLMing}
\alias{LLMing-package}
\title{LLMing: Text Analysis Tools for Psychological Data}
\description{
Package-level documentation and references.
}
\author{
\strong{Maintainer}: Lindley Slipetz \email{ddj6tu@virginia.edu}

}
\keyword{internal}
