% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normahalo.R
\name{normahalo}
\alias{normahalo}
\title{Local outlier score}
\usage{
normahalo(z, rs, S)
}
\arguments{
\item{z}{Dataframe of z scores}

\item{rs}{List of reference sets}

\item{S}{Dataframe of numeric values}
}
\value{
A dataframe of local outlier scores
}
\description{
Computes a normalized Mahalanobis distance score. Only features with nonzero scores in S
receive nonzero Mahalanobis scores.
}
