% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCOS.R
\name{pCOS}
\alias{pCOS}
\title{pCOS scores for every row of dataframe}
\usage{
pCOS(z_dat, vec_SNN)
}
\arguments{
\item{z_dat}{Numeric dataframe, typically z-scores}

\item{vec_SNN}{Numeric dataframe, typically the output of vector_SNN}
}
\value{
A dataframe with same dimensions as z_dat
}
\description{
Applies pCOS_row() to corresponding rows of two data frames,
returning one pCOS value per row.
}
