% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCOS_row.R
\name{pCOS_row}
\alias{pCOS_row}
\title{Pairwise cosine-style row score}
\usage{
pCOS_row(z, v_SNN)
}
\arguments{
\item{z}{Numeric vector}

\item{v_SNN}{Numeric vector, same size as z}
}
\value{
A numeric vector
}
\description{
Given two numeric vectors, computes an average
cosine-based similarity.
}
