% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_set.R
\name{rep_set}
\alias{rep_set}
\title{The vectors of the shared nearest neighbors}
\usage{
rep_set(z, snn)
}
\arguments{
\item{z}{Dataframe of values of reference set}

\item{snn}{Dataframe of shared nearest neighbors indices}
}
\value{
A list of dataframes where each row of the dataframe is the vector representation of a given shared nearest
neighbor
}
\description{
Creates a list of the vectors of the top shared nearest neighbors for each row of the z dataframe
}
