% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_SNN.R
\name{sim_SNN}
\alias{sim_SNN}
\title{Compute shared nearest neighbors}
\usage{
sim_SNN(z_dat, k, tops)
}
\arguments{
\item{z_dat}{A dataframe with numeric columns}

\item{k}{An integer representing number of nearest neighbors}

\item{tops}{An integer representing how many of shared nearest neighbors to return}
}
\value{
A dataframe of top rows with shared nearest neighbors
}
\description{
Builds a shared nearest neighbors matrix and,
for each row (observation), returns the indices of the top neighbors
with the largest SNN overlap counts
}
