% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textanomaly.R
\name{textanomaly}
\alias{textanomaly}
\title{Text anomaly score}
\usage{
textanomaly(dat, k, tops, theta)
}
\arguments{
\item{dat}{A dataframe with text data, one text per row}

\item{k}{An integer representing number of nearest neighbors}

\item{tops}{An integer representing how many of shared nearest neighbors to return}

\item{theta}{Numeric threshold}
}
\value{
Dataframe of local outlier score
}
\description{
Text anomaly detection method adapted from \insertCite{zhang_angle-based_2015}{LLMing}.
}
\references{
\insertRef{zhang_angle-based_2015}{LLMing}
}
