% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_SNN.R
\name{vector_SNN}
\alias{vector_SNN}
\title{Aggregate dataframe into mean feature vectors}
\usage{
vector_SNN(z, snn)
}
\arguments{
\item{z}{Numeric dataframe}

\item{snn}{Dataframe of shared nearest neighbors indices}
}
\value{
Dataframe of same dimensions as z
}
\description{
For each row of the SNN index matrix, this function takes the rows of
reference dataframe, z, and computes their column means, yielding one
mean vector per observation.
}
