% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{Z-score on columns}
\usage{
z_score(dat)
}
\arguments{
\item{dat}{A dataframe with numeric cells}
}
\value{
A dataframe with numeric cells with the same dimensions as dat
}
\description{
Z-score on columns
}
